scripts/mmis_script_test/requirements.txt
requests==2.26.0
google-cloud-storage==1.42.3
